// frontend/src/api/sep_drive/model/inboundChatAction.ts
/**
 * SEP-Drive Gruppe P API
 *
 * Contact: mats.toncik@tnck.dev
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Represents a chat action.
 */
export interface InboundChatAction { 
    /**
     * The type of the message.
     */
    actionType: string;
    /**
     * The content of the message. Might be null if action type is DELETE.
     */
    content?: string;
    /**
     * The id of the chat message. Might be null if action type is SEND.
     */
    chatMessageId?: number;
}