// frontend/src/api/sep_drive/model/outboundChatMessage.ts
/**
 * SEP-Drive Gruppe P API
 *
 * Contact: mats.toncik@tnck.dev
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Represents an outbound chat message.
 */
export interface OutboundChatMessage { 
    /**
     * The id of the trip offer.
     */
    tripOfferId: number;
    /**
     * The id of the chat message.
     */
    chatMessageId: number;
    /**
     * The direction of the message. Either C2D or D2C.
     */
    direction: string;
    /**
     * The creation timestamp of the message.
     */
    createdTimestamp: string;
    /**
     * The latest timestamp of the message.
     */
    timestamp: string;
    /**
     * The content of the message.
     */
    content: string;
    /**
     * Whether the message has been seen by the recipient.
     */
    seen: boolean;
    /**
     * Whether the message has been edited by the sender.
     */
    edited: boolean;
    /**
     * Whether the message has been deleted by the sender.
     */
    deleted: boolean;
}